refactor: extract functions to a new module
