+++ crates/rpc/rpc.rs
@@ -1,10 +1,6 @@
-use std::{
-    future::IntoFuture,
-    net::SocketAddr,
-    time::{SystemTime, UNIX_EPOCH},
-};
-
+use crate::authentication::{validate_jwt_authentication, AuthenticationError};
 use bytes::Bytes;
+use std::{future::IntoFuture, net::SocketAddr};
 
 use axum::{routing::post, Json, Router};
 use axum_extra::{
@@ -22,14 +18,12 @@ use eth::{
     },
     client,
 };
-use hex;
-use jsonwebtoken::{decode, Algorithm, DecodingKey, TokenData, Validation};
-use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use tokio::net::TcpListener;
 use tracing::info;
 use utils::{RpcErr, RpcErrorMetadata, RpcErrorResponse, RpcRequest, RpcSuccessResponse};
 mod admin;
+mod authentication;
 mod engine;
 mod eth;
 mod types;
@@ -42,14 +36,6 @@ use ethereum_rust_storage::Store;
 pub struct ServiceContext {
     storage: Store,
     jwt_secret: Bytes,
-}
-
-// JWT claims struct
-#[derive(Debug, Serialize, Deserialize)]
-struct Claims {
-    iat: usize,
-    id: Option<String>,
-    clv: Option<String>,
 }
 
 pub async fn start_api(
@@ -128,7 +114,6 @@ pub async fn handle_authrpc_request(
                 "id": null
             }))
         }
-
         Err(AuthenticationError::TokenDecodingError) => {
             return Json(json!({
                 "jsonrpc": "2.0",
@@ -139,7 +124,6 @@ pub async fn handle_authrpc_request(
                 "id": null
             }));
         }
-
         Ok(()) => {
             // Proceed with the request
             let req: RpcRequest = serde_json::from_str(&body).unwrap();
@@ -150,36 +134,6 @@ pub async fn handle_authrpc_request(
             rpc_response(req.id, res)
         }
     }
-}
-enum AuthenticationError {
-    InvalidIssuedAtClaim,
-    TokenDecodingError,
-}
-
-fn validate_jwt_authentication(token: &str, secret: Bytes) -> Result<(), AuthenticationError> {
-    let decoding_key = DecodingKey::from_secret(&secret);
-    let mut validation = Validation::new(Algorithm::HS256);
-    validation.validate_exp = false;
-    validation.set_required_spec_claims(&["iat"]);
-    match decode::<Claims>(token, &decoding_key, &validation) {
-        Ok(token_data) => {
-            if invalid_issued_at_claim(token_data) {
-                Err(AuthenticationError::InvalidIssuedAtClaim)
-            } else {
-                Ok(())
-            }
-        }
-        Err(_) => Err(AuthenticationError::TokenDecodingError),
-    }
-}
-
-/// Checks that the "iat" timestamp in the claim is less than 60 seconds from now
-fn invalid_issued_at_claim(token_data: TokenData<Claims>) -> bool {
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_secs() as usize;
-    (now as isize - token_data.claims.iat as isize).abs() > 60
 }
 
 /// Handle requests that can come from either clients or other users
